backend/requirements.txt
amqp==5.1.1
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
confusable-homoglyphs==3.2.0
cryptography==41.0.2
decorator==5.1.1
Django==4.2.4
django-choices==2.0.0
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-encrypted-model-fields==0.6.5
django-js-asset==2.1.0
django-registration==3.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
executing==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipython==8.14.0
jedi==0.19.0
kombu==5.3.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
python-dateutil==2.8.2
python-telegram-bot==20.4
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
stripe==5.4.0
traitlets==5.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
